Separated Cursor.
